requirements.txt
networkx==3.1
numpy==1.24.3
PyYAML==6.0.1
scikit_learn==1.3.2
scipy==1.12.0
torch==1.12.1
torch_geometric==2.4.0
torch_scatter==2.1.0+pt112cu113
tqdm==4.66.1
